Add functionality for deposit moneyand worked with fraction number

+++ SuperBank/Transaction.cs
@@ -3,11 +3,11 @@ namespace SuperBank
 {
     public class Transaction
     {
-        public decimal Amount { get; }
+        public double Amount { get; }
         public DateTime Date { get; }
         public string Notes { get; }
 
-        public Transaction(decimal amount, DateTime date, string note)
+        public Transaction(double amount, DateTime date, string note)
         {
             Amount = amount;
             Date = date;
